core/cache: fix -Wsign-compare warning

+++ core/cache/cache.h
@@ -30,8 +30,8 @@ class Cache {
     typedef CacheIterator<K, V> iterator;
 
     /// Constructor: create a cache with size n.
-    Cache (int n) : capacity (n) {
-      for (int i = 0; i < capacity; ++i){
+    Cache (size_t n) : capacity (n) {
+      for (size_t i = 0; i < capacity; ++i){
         free_list.insert (i);
       }
     }
@@ -63,7 +63,7 @@ class Cache {
 
 
     /// Return number of elements in the cache (Does not work!?).
-    int size() {
+    size_t size() {
       return storage.size();
     }
 
@@ -79,7 +79,7 @@ class Cache {
       std::cout << "cache: add " << key << " ";
 #endif
       if (free_list.size() == 0) {
-        int to_delete = usage[usage.size() - 1];
+        size_t to_delete = usage[usage.size() - 1];
 #ifdef DEBUG_CACHE
         std::cout << "no free space ";
         std::cout << "usage size " << usage.size() << " to_delete=" << to_delete << " key=" << storage[to_delete].first;
@@ -88,7 +88,7 @@ class Cache {
         free_list.insert (to_delete);
       }
 
-      int free_ind = *(free_list.begin());
+      size_t free_ind = *(free_list.begin());
       free_list.erase (free_list.begin());
 #ifdef DEBUG_CACHE
       std::cout << std::endl;
@@ -106,7 +106,7 @@ class Cache {
 
 #ifdef DEBUG_CACHE
       std::cout << "cache usage:";
-      for (int i = 0; i < usage.size(); ++i) {
+      for (size_t i = 0; i < usage.size(); ++i) {
         std::cout << " " << usage[i];
       }
       std::cout << std::endl;
@@ -123,7 +123,7 @@ class Cache {
     /// Get element from the cache.
     V & get (K const & key) {
       assert(contains(key));
-      int ind = index[key];
+      size_t ind = index[key];
 #ifdef DEBUG_CACHE_GET
       std::cout << "cache get: " << key << " ind: " << ind << std::endl;
 #endif
@@ -133,10 +133,10 @@ class Cache {
 
     /// Remove an element from the cache.
     void erase(K const & key) {
-      int i = index[key];
+      size_t i = index[key];
       index.erase(key);
       free_list.insert(i);
-      for (int k = 0; k < usage.size(); ++k) {
+      for (size_t k = 0; k < usage.size(); ++k) {
         if (usage[k] == i) {
           usage.erase(usage.begin() + k);
           break;
@@ -145,13 +145,13 @@ class Cache {
     }
 
     /// Return size of empty space.
-    int space_remains () {
+    size_t space_remains () {
       return free_list.size();
     }
 
     /// Clear the cache.
     void clear () {
-      for (int i = 0; i < capacity; ++i) free_list.insert (i);
+      for (size_t i = 0; i < capacity; ++i) free_list.insert (i);
       index.clear();
       usage.clear();
     }
@@ -183,11 +183,11 @@ class Cache {
 
 private:
 
-    int capacity;
+    size_t capacity;
     std::vector<std::pair<K,V> > storage;
-    std::map<K, int> index;
-    std::set<int> free_list;
-    std::vector<int> usage;
+    std::map<K, size_t> index;
+    std::set<size_t> free_list;
+    std::vector<size_t> usage;
 
     friend class CacheIterator<K, V>;
 
@@ -202,8 +202,8 @@ private:
       }
     }
 
-    void use (int ind) {
-      int i;
+    void use (size_t ind) {
+      size_t i;
       for (i = 0; i < usage.size() && usage[i] != ind; ++i);
       if (i == usage.size()) {
         usage.resize (usage.size()+1);
@@ -221,7 +221,7 @@ template <typename K, typename V>
 std::ostream & operator<< (std::ostream & s, const Cache<K,V> & cache)
 {
   s   << "Cache(\n";
-  for (int i=0; i<cache.storage.size(); i++){
+  for (size_t i=0; i<cache.storage.size(); i++){
     s << "  " << cache.storage[i].first << " => " << cache.storage[i].second << "\n";
   }
   s << ")";
@@ -230,21 +230,21 @@ std::ostream & operator<< (std::ostream & s, const Cache<K,V> & cache)
 
 /// Iterator class for cache
 template <typename K, typename V>
-class CacheIterator : public std::map<K, int>::const_iterator {
+class CacheIterator : public std::map<K, size_t>::const_iterator {
  public:
     std::pair<K, V>& operator*() {
-      return cache->storage[std::map<K, int>::const_iterator::operator*().second];
+      return cache->storage[std::map<K, size_t>::const_iterator::operator*().second];
     }
 
     std::pair<K, V>* operator->() {
-      return &(cache->storage[std::map<K, int>::const_iterator::operator*().second]);
+      return &(cache->storage[std::map<K, size_t>::const_iterator::operator*().second]);
     }
 
  private:
     friend class Cache<K, V>;
     CacheIterator (Cache<K, V>* cache_,
-           typename std::map<K, int>::const_iterator iter_)
-      : std::map<K, int>::const_iterator(iter_),
+           typename std::map<K, size_t>::const_iterator iter_)
+      : std::map<K, size_t>::const_iterator(iter_),
         cache(cache_)
     { }
